Final updates to new Q-Tube 2.0 release
settingsSlice.ts renamed to persistentSlice, it is only used to store persistent data.

reset button no longer changes video type from "playlist" to "video".

Subscriptions are now stored as list of objects that have the name of the user and the name of the channel that is followed.

When the application starts in App.tsx the subscription list is filtered by the user's name, so they only see channels by the name they have subscribed to.

Edits to QuickMythril's bounty commits (showing home page stats and publish video form displaying supported codecs) to improve visibility and readability.